+++ openproblems/tasks/_batch_integration/_common/methods/baseline.py
@@ -1,10 +1,50 @@
 from .....tools.decorators import method
 from .....tools.utils import check_version
-from ...batch_integration_graph.methods.baseline import _randomize_features
-from ...batch_integration_graph.methods.baseline import _randomize_graph
-from ...batch_integration_graph.methods.baseline import _set_uns
 
 import functools
+import numpy as np
+
+
+def _set_uns(adata):
+    adata.uns["neighbors"] = adata.uns["uni"]
+    adata.uns["neighbors"]["connectivities_key"] = "connectivities"
+    adata.uns["neighbors"]["distances_key"] = "distances"
+
+
+def _randomize_features(X, partition=None):
+    X_out = X.copy()
+    if partition is None:
+        partition = np.full(X.shape[0], 0)
+    else:
+        partition = np.asarray(partition)
+    for partition_name in np.unique(partition):
+        partition_idx = np.argwhere(partition == partition_name).flatten()
+        X_out[partition_idx] = X[np.random.permutation(partition_idx)]
+    return X_out
+
+
+def _randomize_graph(adata, partition=None):
+    distances, connectivities = (
+        adata.obsp["uni_distances"],
+        adata.obsp["uni_connectivities"],
+    )
+    new_idx = _randomize_features(np.arange(distances.shape[0]), partition=partition)
+    adata.obsp["distances"] = distances[new_idx][:, new_idx]
+    adata.obsp["connectivities"] = connectivities[new_idx][:, new_idx]
+    _set_uns(adata)
+    return adata
+
+
+def _random_embedding(partition):
+    from sklearn.preprocessing import LabelEncoder
+    from sklearn.preprocessing import OneHotEncoder
+
+    embedding = OneHotEncoder().fit_transform(
+        LabelEncoder().fit_transform(partition)[:, None]
+    )
+    embedding = embedding + np.random.uniform(-0.01, 0.01, embedding.shape)
+    return embedding
+
 
 _baseline_method = functools.partial(
     method,
