+++ openproblems/tasks/_batch_integration/batch_integration_embed/methods/baseline.py
@@ -1,7 +1,6 @@
 from .....tools.decorators import method
 from .....tools.utils import check_version
-from ...batch_integration_graph.methods.baseline import _random_embedding
-from ...batch_integration_graph.methods.baseline import _randomize_features
+from ..._common.methods.baseline import _random_embedding
 
 import functools
 import numpy as np
@@ -18,55 +17,10 @@ _baseline_method = functools.partial(
 
 
 @_baseline_method(
-    method_name="No Integration",
-)
-def no_integration(adata, test=False):
-    adata.obsm["X_emb"] = adata.obsm["X_uni_pca"]
-    adata.uns["method_code_version"] = check_version("openproblems")
-    return adata
-
-
-@_baseline_method(
-    method_name="Random Integration",
-)
-def random_integration(adata, test=False):
-    adata.obsm["X_emb"] = _randomize_features(adata.obsm["X_uni_pca"])
-    adata.uns["method_code_version"] = check_version("openproblems")
-    return adata
-
-
-@method(
-    method_name="Random Integration by Celltype",
-    paper_name="Random Integration by Celltype (baseline)",
-    paper_reference="openproblems",
-    paper_year=2022,
-    code_url="https://github.com/openproblems-bio/openproblems",
-    is_baseline=True,
-)
-def celltype_random_integration(adata, test=False):
-    adata.obsm["X_emb"] = _randomize_features(
-        adata.obsm["X_uni_pca"], partition=adata.obs["labels"]
-    )
-    adata.uns["method_code_version"] = check_version("openproblems")
-    return adata
-
-
-@_baseline_method(
     method_name="Random Embedding by Celltype",
 )
 def celltype_random_embedding(adata, test=False):
     adata.obsm["X_emb"] = _random_embedding(partition=adata.obs["labels"])
-    adata.uns["method_code_version"] = check_version("openproblems")
-    return adata
-
-
-@_baseline_method(
-    method_name="Random Integration by Batch",
-)
-def batch_random_integration(adata, test=False):
-    adata.obsm["X_emb"] = _randomize_features(
-        adata.obsm["X_uni_pca"], partition=adata.obs["batch"]
-    )
     adata.uns["method_code_version"] = check_version("openproblems")
     return adata
