+++ openproblems/tasks/_batch_integration/batch_integration_graph/methods/__init__.py
@@ -1,7 +1,7 @@
-from .baseline import batch_random_integration
+from ..._common.methods.baseline import batch_random_integration
 from .baseline import celltype_random_graph
-from .baseline import celltype_random_integration
-from .baseline import no_integration
+from ..._common.methods.baseline import celltype_random_integration
+from ..._common.methods.baseline import no_integration
 from .baseline import random_integration
 from .bbknn import bbknn_full_scaled
 from .bbknn import bbknn_full_unscaled
